+++ TesterCache/CacheEntry.cs
@@ -5,7 +5,7 @@ using DeltaSnapshot;
 namespace TesterCache {
     public class CacheEntry<TEntity> where TEntity : class, IDataSetEntity, new() {
         public long CacheSnapshotId { get; set; }
-        public int SubscriptionDataSetId { get; set; }
+        public long SubscriptionDataSetId { get; set; }
         public long RunId { get; set; }
         public string EntityIdentifier { get; set; }
         public string EntityDeltaCode { get; set; }
